+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file lists all changes to this project, grouped by versions that follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html). This file is based on the format set forward by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- Security headers with Flask-Talisman, including HSTS
+- Custom Content-Security-Policy header that allows unsafe-inline style-src
+  and connect-src for relevant ws:// and wss:// headers
+
 ## [0.6.2] - 2020-03-10
 
 ### Added